Move `SdfAssetPath` tests to their own files
